Add deserialization to Serialize
Remove the old Serializable class decorator, and
use the Serialize mixin constructor everywhere.
Add deserialization to Serialize, although it doesn't
quite work yet: we want to be able to recursively descend
through each field and deserialize it too, but we need to
give Python the type annotations in order to do that, and
it is fussing about optionally giving a type.
Hopefully, this is fixable!