Add many more AST comps, handle more from_dict
Upgrade from_dict to be able to handle lists and dicts.
Clean up the doctest and also add a protocol Serializable
to make it easy to check for stuff we can serialize.

Add a whole whack of new classes to the AST module (now aast.py)
to handle many more of the elements of the JSON output.