Add a bunch more dataclasses, fix bug in from_dict
Fix a bug in from_dict when handling lists.
Some open questions to consider when handling deserialization
involve using the class or type key of the JSON dict to narrow
the type of the returned dict: this is necessary when we have
a class that can store subclasses in one of its fields, e.g.
TraceableStatement can store multiple kinds of Statement.
This might be doable by adding an argument to Serialize to
narrow the type using a particular field.