Fix MRO issue with Serialize
Serialize has issues with MRO when included after a base class: if the
base class also has Serialize as a mixin, then its definitions will
shadow the intended Serialize mixin for this class. This is intended
behavior, as described here:
https://nedbatchelder.com/blog/201210/multiple_inheritance_is_hard.html

To address this, we need to place Serialize ahead of any base classes in
derived classes. This commit does so and appears to work without causing
any issues except for a warning from pyright about a lack of consistent
method ordering. We retain some unused code attempting to make a class
decorator Serialize3 which was intended to try and resolve that issue,
but can probably be removed if the typing warning can be hidden or
resolved.