Change to use SerializeMixin over Serialize
Change all uses of Serialize to SerializeMixin in bat_ast.py
Requires renaming some arguments due to parameter conflict for "name".
One potential issue we can foresee with SerializeMixin is if multiple
classes with __init_subclass__ are used: as SerializeMixin takes all
keyword arguments and has to go at the end, we can imagine an issue
where it consumes arguments intended for another class.
It's not totally clear to me how that might take place, but one way to
preempt that issue would be to define SerializeMixin to take a single
argument fields, which contains the dict, rather than using **kwargs.