Add Field for Serialize, allow defaults
Add a Field class to Serialize to give more details for the fields
generated. A Field has a name and type as before, but can also define a
default value, which is used when the given json_name is not found in
the dictionary. This value will supersede any defaults declared over the
actual class instance: it is passed to the constructor as though a
value was actually specified, meaning it can generate instances which
differ from those constructed with the default given for the __init__
function.