Add some initial code for converting bast to aast
Add a to_aast method to bast.base.ASTNode to convert from bast to aast.
Also add an (unused) class decorator converter.ToAast to help with this
process: the idea of the decorator is to allow us to save writing
boilerplate and automatically provide a to_aast method to the basic bast
expressions that look the same in aast.