Various corrections to bast, esp communities
Fix various parts of bast which incorrectly modelled the Batfish
behavior, esp. for communities.
We now distinguish CommunitySetExprs (a set of literal communities),
CommunityMatchExprs (a match condition on a single community) and
CommunitySetMatchExprs (a match condition on a set of communities:
generally appearing to be a disjunction over CommunityMatchExprs).
Also separate PrefixSetExpr from PrefixExpr, and move non-boolean
community expressions into communities.
Change parse_bf_clsname to be able to handle unqualified class names as
well.