Change aast.Expression to be generic, flatten
Move all aast expressions under Expression, and make Expression generic.
The idea behind doing this is to make better use of the type system when
constructing expressions and avoiding the need for casting.
All expressions are now parameterized by a type, which ensures that they
are put together correctly at the call site.