Add SkipStatement (no-op) and convert_stmts fn
Add a SkipStatement which represents a no-op, useful for when an
IfStatement or SeqStatement is followed by an operation we don't want to
model.
Change type signature for convert_stmt to just return a single
statement, and add a convert_stmts that can convert a list of Batfish
statements into a sequence of Angler statements connected by
SeqStatement.
Correct bug in type of bast.topology.Node.