Add explicit types for exprs and stmts
Add explicit types for exprs and stmts to ensure that a user of Angler
can determine the precise type of every term without needing to do type
inference. This simplifies deserialization significantly.
The tricky part is that our conversion code needs to ensure that these
types are annotated correctly: right now, we're basically just labelling
things and hoping they're correct, but it's possible that there may be
bugs or mistakes in places we'd have to address.
The current implementation is also kind of messy: a better approach
would be to have a $type field we serialize, and a ty_args tuple we pass
to each expr and stmt that requires it, such that $type can be
serialized without ever needing to be modified by the user/constructor,
whereas ty_args will depend on where the expr or stmt is being used.