Fix code for generating RoutingPolicy w a bind
Add the equivalent of a bind operation over the current RoutingPolicy
code, transforming statements from Route to a Pair(Bool, Route) to
statements from Pair(Bool, Route) to Pair(Bool, Route).
There are still some inconsistencies with how we label types, and the
infrastructure for giving annotations needs further simplification as
well as corrections for the inference of types for IfStatements, but
it's looking a little less wonky now.

Also add JSON serialization for TypeAnnotations, and make minor changes
to naming of exported aast.prog.Program fields.