Add more prefixes during convert, try to fix edges
Add the interface IP addresses to the prefixes when converting
(might not be correct to do if they're supposed to be internal).
Also store prefixes in a set for now just to avoid duplicate additions.
We still seem to not add all edges present in the network due to how our
code is currently written: this needs to be fixed before angler output
is usable.
It looks like it may be that the ordering of the edges affects which
ones make it in or not.