Add and update various Batfish exprs
Break apart handling of routes with a single boolean into a Result type.
(We'll probably eventually merge this into the route, rather than using
a Pair).
The Result type is meant to distinguish the differences between exiting
vs. returning and other conditions that arise when handling some of
Juniper's policy chaining behaviour.