Update properties, add BigInt
Add unfinished code to convert Batfish routing policies into Angler
functions, such that the special early-return behavior of policy
statements is properly captured.
We need to add logic to check that, if a statement returns, then
subsequent statements should not be executed: this can be done by
nesting the subsequent statements inside a newly-added if statement.
The resulting policy behavior should then ensure that all control flow
is characterized by the if statement conditions.