Remove query functionality, add Poetry
Remove the querying functionality from angler.
This functionality blurred the lines of what angler is meant to do: we
want to let end-users write systems that can query the network's
behavior using angler's output, not have angler define the queries for
them.
Doing this simplifies the angler "Program" into simply a representation
of the network. There are still changes we can make, e.g. to replace the
symbolics field with a list of external neighbors. That said, the
Program can now be reframed as a Network class which simply defines the
network's components (essentially just the nodes, route type and
transfer functions).

The second big change in this commit is to introduce Poetry as our build
system. This should make it simpler to manage dependencies, at the cost
of (of course) having a new dependency to install to manage them.
This commit also shows some use of pdoc, as a doc dependency,
to document the API accordingly.